requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
pillow>=10.0.0
insightface>=0.7.3
opencv-python>=4.8.0
numpy>=1.24.0
python-multipart>=0.0.9
sqlmodel>=0.0.14
pytest>=7.4.0
httpx>=0.26.0
python-dotenv>=1.0.0
streamlit
requests